src/multi-platform/caller/requirements.txt
requests>=2.31.0
opentelemetry-distro>=0.45b0
opentelemetry-exporter-otlp>=1.24.0
opentelemetry-propagator-aws-xray>=1.0.0
opentelemetry-instrumentation-requests>=0.45b0
